+++ src/main/java/pikabot/Parser.java
@@ -1,3 +1,23 @@
+package pikabot;
+
+import pikabot.command.Command;
+import pikabot.command.DeadlineCommand;
+import pikabot.command.DeleteCommand;
+import pikabot.command.EventCommand;
+import pikabot.command.InvalidCommand;
+import pikabot.command.ListCommand;
+import pikabot.command.MarkCommand;
+import pikabot.command.TodoCommand;
+import pikabot.command.UnmarkCommand;
+
+import pikabot.task.Deadline;
+import pikabot.task.Event;
+import pikabot.task.Todo;
+
+import pikabot.exception.DeadlineException;
+import pikabot.exception.EventException;
+import pikabot.exception.TodoException;
+
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
