+++ src/main/java/pikabot/TaskList.java
@@ -1,3 +1,7 @@
+package pikabot;
+
+import pikabot.task.Task;
+
 import java.util.ArrayList;
 
 public class TaskList {
