+++ src/main/java/pikabot/command/Command.java
@@ -1,4 +1,7 @@
-import java.io.IOException;
+package pikabot.command;
+
+import pikabot.TaskList;
+import pikabot.Storage;
 
 public abstract class Command {
     public abstract void execute(TaskList taskList, Storage storage);
