+++ src/main/java/pikabot/command/EventCommand.java
@@ -1,3 +1,13 @@
+package pikabot.command;
+
+import pikabot.TaskList;
+import pikabot.Storage;
+import pikabot.Parser;
+import pikabot.Ui;
+
+import pikabot.task.Event;
+import pikabot.exception.EventException;
+
 import java.io.IOException;
 import java.time.format.DateTimeParseException;
