+++ src/main/java/pikabot/exception/DeadlineException.java
@@ -1,5 +1,6 @@
-public class DeadlineException extends DukeException {
+package pikabot.exception;
 
+public class DeadlineException extends PikaBotException {
     public DeadlineException(String message) {
         super("☹ OOPS!!! " + message);
     }
