+++ src/main/java/pikabot/exception/EventException.java
@@ -1,5 +1,6 @@
-public class EventException extends DukeException {
+package pikabot.exception;
 
+public class EventException extends PikaBotException {
     public EventException(String message) {
         super("☹ OOPS!!! " + message);
     }
