+++ src/main/java/pikabot/exception/InvalidTaskCommandException.java
@@ -1,5 +1,6 @@
-public class InvalidTaskCommandException extends DukeException {
+package pikabot.exception;
 
+public class InvalidTaskCommandException extends PikaBotException {
     public InvalidTaskCommandException() {
         super("I'm sorry, but I don't know what that means :-(");
     }
