+++ src/main/java/pikabot/exception/PikaBotException.java
@@ -1,0 +1,7 @@
+package pikabot.exception;
+
+public class PikaBotException extends Exception {
+    public PikaBotException(String message) {
+        super(message);
+    }
+}
