+++ src/main/java/pikabot/exception/TodoException.java
@@ -1,5 +1,6 @@
-public class TodoException extends DukeException {
+package pikabot.exception;
 
+public class TodoException extends PikaBotException {
     public TodoException() {
         super("☹ OOPS!!! The description of a todo cannot be empty.");
     }
