+++ src/main/java/pikabot/task/Deadline.java
@@ -1,3 +1,5 @@
+package pikabot.task;
+
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
