+++ src/main/java/pikabot/task/Event.java
@@ -1,7 +1,7 @@
+package pikabot.task;
+
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
-
-
 
 public class Event extends Task {
