+++ src/main/java/pikabot/task/Task.java
@@ -1,3 +1,4 @@
+package pikabot.task;
 public class Task {
     protected String description;
     protected boolean isDone;
