+++ src/main/java/pikabot/task/Todo.java
@@ -1,5 +1,6 @@
-public class Todo extends Task {
+package pikabot.task;
 
+public class Todo extends Task {
     public Todo(String description) {
         super(description);
     }
